Move edit and delete dialog to different components
